Add support for PlantUML rendering in Neovim
This commit adds a new Lua file, `plantuml.lua`, which enables the rendering of PlantUML diagrams in Neovim. It includes a setup function that configures the renderer to use the 'image' type and sets the program to 'feh'. Additionally, it specifies a dependency on the 'aklt/plantuml-syntax' plugin and sets `render_on_write` to true.

Link: https://gitlab.com/itaranto/plantuml.nvim